Do not handle PipelineCancelledExceptions
PipelineCancelledExceptions occur when the pipeline was cancelled by intention. This happens when a request was aborted and no response is expected anymore. The PipelineCancelledException is thrown to make sure the pipeline and all it steps will actually get a call-back and get into a "final" state to prevent resource leaks etc.

Signed-off-by: Benjamin Rögner <benjamin.roegner@here.com>