DS-1037: MapGIT - Add implementation for branching of spaces
A branch is like a separate space, which is based on a specific version of the space
or on a specific version of some other branch of the space. Reading & writing features
to a branch does not have an effect on the main branch or any other branch of the space.

A branch can be created by pointing to a base ref. A base ref consists of the base branch
(e.g, "main") and the base version of that base branch. New write transactions to the
created branch will continue to increase the version counter starting from the
base version. Versions in that branch are independent from the versions of the base branch.

Also implemented the following operations for now:

- Merge (to merge the content / changes from one branch into another one)
- Rebase (to change the base version of a branch and base all changes on top of that new base version)

Signed-off-by: Benjamin Rögner <benjamin.roegner@here.com>