DS-1101: Implement named user input-sets for jobs
This change enables the user to define multiple input sets that are distinguishable by names. Depending on the expectations of the job, the different input sets can be assigned to be used for different purposes by the job compiler.

The change is backwards compatible. If the user provides an "unnamed" input-set it has the default name "inputs".

Signed-off-by: Benjamin Rögner <benjamin.roegner@here.com>