DS-1184: Add BWC-workaround for missing base versions on feature creations
If the base version was not defined by the user, but a version conflict detection was requested, the FeatureWriter was throwing an exception until now. While that behavior is basically correct, this workaround is necessary to keep the backwards compatibility temporarily. That workaround can be removed again once the global base version can be set by the user and the "conflictDetection" param was deprecated

Signed-off-by: Benjamin Rögner <benjamin.roegner@here.com>